Refactor polling and executive pages to integrate legacy and current governance systems. Update imports and types for better structure. Add alerts for governance transitions and improve user navigation between legacy and new systems.
